+++ backend/src/akvo/lumen/endpoint/transformation.clj
@@ -1,8 +1,7 @@
 (ns akvo.lumen.endpoint.transformation
-  (:require [compojure.core :refer :all]
-            [akvo.lumen.component.tenant-manager :refer [connection]]
+  (:require [akvo.lumen.component.tenant-manager :refer [connection]]
             [akvo.lumen.transformation :as t]
-            [ring.util.response :as response]))
+            [compojure.core :refer :all]))
 
 (defn endpoint [{:keys [tenant-manager transformation-engine]}]
   (context "/api/transformations" {:keys [tenant] :as request}
@@ -10,16 +9,14 @@
       (context "/:dataset-id" [dataset-id]
 
         (POST "/transform" {:keys [body] :as request}
-          (merge (response/response {})
-                 (t/schedule tenant-conn
-                             transformation-engine
-                             dataset-id
-                             {:type :transformation
-                              :transformation body})))
+          (t/schedule tenant-conn
+                      transformation-engine
+                      dataset-id
+                      {:type :transformation
+                       :transformation body}))
 
         (POST "/undo" _
-          (merge (response/response {})
-                 (t/schedule tenant-conn
-                             transformation-engine
-                             dataset-id
-                             {:type :undo})))))))
+          (t/schedule tenant-conn
+                      transformation-engine
+                      dataset-id
+                      {:type :undo}))))))