+++ backend/test/akvo/lumen/endpoint/collection_test.clj
@@ -5,10 +5,12 @@
                                          migrate-tenant
                                          rollback-tenant]]
             [akvo.lumen.import.csv-test :refer [import-file]]
+            [akvo.lumen.lib :as lib]
             [akvo.lumen.lib.collection :as collection]
             [akvo.lumen.lib.dashboard :as dashboard]
             [akvo.lumen.lib.dataset :as dataset]
             [akvo.lumen.lib.visualisation :as visualisation]
+            [akvo.lumen.variant :as variant]
             [clojure.test :refer :all]
             [com.stuartsierra.component :as component]
             [duct.component.hikaricp :refer [hikaricp]]))
@@ -43,11 +45,13 @@
 
 (defn create-visualisation [tenant-conn dataset-id]
   (-> (visualisation/create *tenant-conn* (visualisation-body dataset-id) {})
-      (get-in [:body "id"])))
+      variant/value
+      (get "id")))
 
 (defn create-dashboard [tenant-conn]
   (-> (dashboard/create tenant-conn {"title" ""})
-      (get-in [:body :id])))
+      variant/value
+      (get :id)))
 
 (deftest ^:functional collection-test
   (let [;; Import a few datasets
@@ -58,52 +62,53 @@
         db1 (create-dashboard *tenant-conn*)
         db2 (create-dashboard *tenant-conn*)]
     (testing "Create an empty collection"
-      (let [response (collection/create *tenant-conn* {"title" "col1"})]
-        (is (= 201 (:status response)))
+      (let [[tag collection] (collection/create *tenant-conn* {"title" "col1"})]
+        (is (= ::lib/created tag))
         (is (= #{:id :title :modified :created :entities}
-               (-> response :body keys set)))
-        (is (= 409 (:status (collection/create *tenant-conn* {"title" "col1"}))))
-        (is (= 400 (:status (collection/create *tenant-conn* {"title" nil}))))
-        (is (= 400 (:status (collection/create *tenant-conn* {"title" (apply str (repeat 129 "a"))}))))))
+               (-> collection keys set)))
+        (is (= ::lib/conflict (first (collection/create *tenant-conn* {"title" "col1"}))))
+        (is (= ::lib/bad-request (first (collection/create *tenant-conn* {"title" nil}))))
+        (is (= ::lib/bad-request (first (collection/create *tenant-conn*
+                                                           {"title" (apply str (repeat 129 "a"))}))))))
     (testing "Create a collection with entities"
-      (let [response (collection/create *tenant-conn* {"title" "col2"
-                                                       "entities" [ds1 vs1 db1]})]
-        (is (= 201 (:status response)))
-        (is (= #{ds1 vs1 db1} (-> response :body :entities set)))))
+      (let [[tag collection] (collection/create *tenant-conn* {"title" "col2"
+                                                               "entities" [ds1 vs1 db1]})]
+        (is (= ::lib/created tag))
+        (is (= #{ds1 vs1 db1} (-> collection :entities set)))))
 
     (testing "Fetch collection"
       (let [id (-> (collection/create *tenant-conn* {"title" "col3" "entities" [ds2 vs2 db2]})
-                   :body :id)
-            coll (-> (collection/fetch *tenant-conn* id) :body)]
+                   variant/value :id)
+            coll (variant/value (collection/fetch *tenant-conn* id))]
         (is (= #{ds2 vs2 db2} (-> coll :entities set)))))
 
     (testing "Update collection"
       (let [id (-> (collection/create *tenant-conn* {"title" "col4" "entities" [ds2 vs2 db2]})
-                   :body :id)]
-        (let [coll (-> (collection/update *tenant-conn* id {"title" "col4 renamed"}) :body)]
+                   variant/value :id)]
+        (let [coll (variant/value (collection/update *tenant-conn* id {"title" "col4 renamed"}))]
           (is (= (:title coll) "col4 renamed"))
           (is (= (-> coll :entities set) #{ds2 vs2 db2}))
           (collection/update *tenant-conn* id {"entities" []})
-          (is (= [] (-> (collection/fetch *tenant-conn* id) :body :entities)))
+          (is (= [] (-> (collection/fetch *tenant-conn* id) variant/value :entities)))
           (collection/update *tenant-conn* id {"entities" [ds1 vs1]})
-          (is (= #{ds1 vs1} (-> (collection/fetch *tenant-conn* id) :body :entities set)))
+          (is (= #{ds1 vs1} (-> (collection/fetch *tenant-conn* id) variant/value :entities set)))
           (collection/update *tenant-conn* id {"entities" [vs1 vs2]})
-          (is (= #{vs1 vs2} (-> (collection/fetch *tenant-conn* id) :body :entities set))))))
+          (is (= #{vs1 vs2} (-> (collection/fetch *tenant-conn* id) variant/value :entities set))))))
 
     (testing "Delete collection"
-      (let [id (-> (collection/create *tenant-conn* {"title" "col5"}) :body :id)]
+      (let [id (-> (collection/create *tenant-conn* {"title" "col5"}) variant/value :id)]
         (collection/delete *tenant-conn* id)
-        (is (= 404 (:status (collection/fetch *tenant-conn* id))))))
+        (is (= ::lib/not-found (variant/tag (collection/fetch *tenant-conn* id))))))
 
     (testing "Delete collection entities"
       (let [id (-> (collection/create *tenant-conn* {"title" "col6"
                                                      "entities" [ds1 ds2 vs1 vs2 db1 db2]})
-                   :body :id)]
+                   variant/value :id)]
         (dashboard/delete *tenant-conn* db1)
         (is (= #{ds1 ds2 vs1 vs2 db2}
                (-> (collection/fetch *tenant-conn* id)
-                   :body :entities set)))
+                   variant/value :entities set)))
         (dataset/delete *tenant-conn* ds1)
         (is (= #{ds2 vs2 db2}
                (-> (collection/fetch *tenant-conn* id)
-                   :body :entities set)))))))
+                   variant/value :entities set)))))))