+++ backend/test/akvo/lumen/lib/pivot_test.clj
@@ -5,6 +5,7 @@
                                          migrate-tenant
                                          rollback-tenant]]
             [akvo.lumen.import.csv-test :refer [import-file]]
+            [akvo.lumen.lib :as lib]
             [akvo.lumen.lib.aggregation :as aggregation]
             [akvo.lumen.transformation :as tf]
             [clojure.test :refer :all]
@@ -48,38 +49,38 @@
 (deftest ^:functional test-pivot
   (let [query (partial aggregation/query *tenant-conn* *dataset-id* "pivot")]
     (testing "Empty query"
-      (let [{:keys [status body]} (query {"aggregation" "count"})]
-        (is (= status 200))
-        (is (= body {:columns [{"type" "number" "title" "Total"}]
-                     :rows [[8]]
-                     :metadata {"categoryColumnTitle" nil}}))))
+      (let [[tag query-result] (query {"aggregation" "count"})]
+        (is (= tag ::lib/ok))
+        (is (= query-result {:columns [{"type" "number" "title" "Total"}]
+                             :rows [[8]]
+                             :metadata {"categoryColumnTitle" nil}}))))
 
     (testing "Empty query with filter"
-      (let [{:keys [status body]} (query {"aggregation" "count"
-                                          "filters" [{"column" "c1"
-                                                      "value" "a1"
-                                                      "operation" "keep"
-                                                      "strategy" "is"}]})]
-        (is (= status 200))
-        (is (= body {:columns [{"type" "number" "title" "Total"}]
-                     :rows [[4]]
-                     :metadata {"categoryColumnTitle" nil}}))))
+      (let [[tag query-result] (query {"aggregation" "count"
+                                       "filters" [{"column" "c1"
+                                                   "value" "a1"
+                                                   "operation" "keep"
+                                                   "strategy" "is"}]})]
+        (is (= tag ::lib/ok))
+        (is (= query-result {:columns [{"type" "number" "title" "Total"}]
+                             :rows [[4]]
+                             :metadata {"categoryColumnTitle" nil}}))))
 
     (testing "Category column only"
-      (let [{:keys [status body]} (query {"aggregation" "count"
-                                          "categoryColumn" "c1"})]
-        (is (= status 200))
-        (is (= body {:columns [{"title" "" "type" "text"}
-                               {"title" "a1" "type" "number"}
-                               {"title" "a2" "type" "number"}]
-                     :rows [["Total" 4 4]]
-                     :metadata {"categoryColumnTitle" "A"}}))))
+      (let [[tag query-result] (query {"aggregation" "count"
+                                       "categoryColumn" "c1"})]
+        (is (= tag ::lib/ok))
+        (is (= query-result {:columns [{"title" "" "type" "text"}
+                                       {"title" "a1" "type" "number"}
+                                       {"title" "a2" "type" "number"}]
+                             :rows [["Total" 4 4]]
+                             :metadata {"categoryColumnTitle" "A"}}))))
 
     (testing "Row Column Only"
-      (let [{:keys [status body]} (query {"aggregation" "count"
-                                          "rowColumn" "c2"})]
-        (is (= status 200))
-        (is (= body
+      (let [[tag query-result] (query {"aggregation" "count"
+                                       "rowColumn" "c2"})]
+        (is (= tag ::lib/ok))
+        (is (= query-result
                {:columns [{"type" "text", "title" "B"}
                           {"type" "number", "title" "Total"}],
                 :rows [["b1" 4]
@@ -87,11 +88,11 @@
                 :metadata {"categoryColumnTitle" nil}}))))
 
     (testing "Row & Category Column with count aggregation"
-      (let [{:keys [status body]} (query {"aggregation" "count"
-                                          "categoryColumn" "c1"
-                                          "rowColumn" "c2"})]
-        (is (= status 200))
-        (is (= body
+      (let [[tag query-result] (query {"aggregation" "count"
+                                       "categoryColumn" "c1"
+                                       "rowColumn" "c2"})]
+        (is (= tag ::lib/ok))
+        (is (= query-result
                {:columns [{"title" "B", "type" "text"}
                           {"title" "a1", "type" "number"}
                           {"title" "a2", "type" "number"}]
@@ -100,12 +101,12 @@
                 :metadata {"categoryColumnTitle" "A"}}))))
 
     (testing "Row & Category Column with mean aggregation"
-      (let [{:keys [status body]} (query {"aggregation" "mean"
-                                          "categoryColumn" "c1"
-                                          "rowColumn" "c2"
-                                          "valueColumn" "c3"})]
-        (is (= status 200))
-        (is (= body
+      (let [[tag query-result] (query {"aggregation" "mean"
+                                       "categoryColumn" "c1"
+                                       "rowColumn" "c2"
+                                       "valueColumn" "c3"})]
+        (is (= tag ::lib/ok))
+        (is (= query-result
                {:columns [{"title" "B", "type" "text"}
                           {"title" "a1", "type" "number"}
                           {"title" "a2", "type" "number"}]
@@ -114,16 +115,16 @@
                 :metadata {"categoryColumnTitle" "A"}}))))
 
     (testing "Row & Category Column with mean aggregation and filter"
-      (let [{:keys [status body]} (query {"aggregation" "mean"
-                                          "categoryColumn" "c1"
-                                          "rowColumn" "c2"
-                                          "valueColumn" "c3"
-                                          "filters" [{"column" "c3"
-                                                      "value" "11"
-                                                      "operation" "remove"
-                                                      "strategy" "isHigher"}]})]
-        (is (= status 200))
-        (is (= body
+      (let [[tag query-result] (query {"aggregation" "mean"
+                                       "categoryColumn" "c1"
+                                       "rowColumn" "c2"
+                                       "valueColumn" "c3"
+                                       "filters" [{"column" "c3"
+                                                   "value" "11"
+                                                   "operation" "remove"
+                                                   "strategy" "isHigher"}]})]
+        (is (= tag ::lib/ok))
+        (is (= query-result
                {:columns [{"title" "B" "type" "text"}
                           {"title" "a1" "type" "number"}
                           {"title" "a2" "type" "number"}]