+++ ci/dev/deploy.sh
@@ -11,4 +11,6 @@ sudo /opt/google-cloud-sdk/bin/gcloud docker -- push eu.gcr.io/${PROJECT_NAME}/l
 
 sudo chown -R ubuntu:ubuntu /home/ubuntu/.kube
 
-kubectl patch deployment lumen -p '{"spec":{"template":{"spec":{"containers":[{"name":"lumen-backend","image":"eu.gcr.io/akvo-lumen/lumen-backend:'"$CIRCLE_SHA1"'"},{"name":"lumen-client","image":"eu.gcr.io/akvo-lumen/lumen-client:'"$CIRCLE_SHA1"'"}]}}}}'
+sed -e "s/\${BUILD_HASH}/$CIRCLE_SHA1/" ci/deployment.yaml.template > deployment.yaml
+
+kubectl apply -f deployment.yaml