+++ RELEASE_NOTES.md
@@ -1,5 +1,13 @@
 # Release Notes
 
+## 0.10.1
+
+### Bugfixes
+
+* Fixed bug where the user would stop being authenticated after keeping lumen open but
+  not using it for 30 minutes
+* Fixed a bug where some columns for flow imported datasets could not be transformed
+
 ## 0.10.0
 
 ### New and noteworthy